+++ .vscode/launch.json
@@ -9,10 +9,8 @@
             "type": "go",
             "request": "launch",
             "mode": "auto",
-            "program": "${workspaceFolder}/lnkshrt",
-            "env": {
-                "LNKSHRT_PERSISTENT": "1",
-            }
+            "program": "${workspaceFolder}/cmd/lnkshrt/main.go",
+            "env": {}
         }
     ]
 }
