+++ go.mod
@@ -6,7 +6,4 @@ require github.com/gorilla/mux v1.8.0
 
 require github.com/mattn/go-sqlite3 v1.14.12
 
-require (
-	github.com/felixge/httpsnoop v1.0.1 // indirect
-	github.com/gorilla/handlers v1.5.1
-)
+require github.com/gorilla/handlers v1.5.1
