--- lnkshrt/.dockerignore
@@ -1,3 +1,0 @@
-Dockerfile
-.dockerignore
-README.md
