+++ scripts/build.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/sh
+go build -v -o ./lnkshrt main.go
