+++ web/lnkshrt-gui/.eslintrc.json
