+++ web/lnkshrt-gui/models/link.ts
