+++ web/lnkshrt-gui/next-env.d.ts
