+++ web/lnkshrt-gui/next.config.js
