+++ web/lnkshrt-gui/public/favicon.ico
