+++ web/lnkshrt-gui/styles/globals.css
