+++ web/lnkshrt-gui/tsconfig.json
